refactor(account-api): refactor account api to return GetAccountPresenter and refactor account controller to return GetAccountPresenter
